pyproject.toml
[project]
name = "rag-router"
version = "0.1.0"
description = "A simplified router for Retrieval-Augmented Generation (RAG)"
readme = "README.md"
requires-python = ">=3.10,<3.12"
dependencies = [
    "fastapi==0.115.12",
    "uvicorn==0.34.2",
    "httpx==0.28.1",
    "pymilvus==2.5.8",
    "tiktoken==0.9.0",
    "pyyaml==6.0.2",
    "llama-stack-client==0.2.10",
    "fire==0.7.1",
    "jinja2==3.1.6",
    "deepeval>=3.5.9",
    "pandas>=2.3.2",
    "tqdm>=4.67.1",
    "rich>=14.1.0",
    "numpy>=2.2.6",
    "scikit-learn>=1.7.2",
    "plotly>=6.3.0",
    "matplotlib>=3.10.6",
    "seaborn>=0.13.2",
]

[project.optional-dependencies]
dev = [
    "black",
    "isort",
    "pytest",
] 
